flask-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
Werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.3.0
requests==2.31.0
celery==5.3.4
redis==5.0.1

# New dependencies for AI and advanced features
google-generativeai==0.3.0
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3
openpyxl==3.1.2